衝突で消えたので再作業 [update]一旦1Fのシステム配置完了 [add]少女のアニメーションを変更(イラスト反映)
