+++ Assets/Animation/GirlDown.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 672fd805311747549ae1d98973434f9d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
