+++ Assets/Animation/GirlLeft.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c2fbbf7bb143e6c49a402ed4cfa740b4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
