+++ Assets/Animation/GirlRight.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4c6e152c191a01d40a0000972749313a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
