+++ Assets/Animation/GirlUp.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ece8313d320705b4dae318d612b92a8e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
