+++ Assets/Image/GirlMotion.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 22de6be3fbf77df4cb452ac801da5120
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
