+++ Assets/Image/GirlMotion/Down.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bede64ddc58ae4046bbc04b375dd52b3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
