+++ Assets/Image/GirlMotion/Left.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2863d0399cfa5c438fed0f87cf231ac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
