+++ Assets/Image/GirlMotion/Right.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 66ed8edcfc32bc245b4d55affa7b0046
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
