+++ Assets/Image/GirlMotion/Up.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 018ae0e830f74164ba07ce77c0b77954
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
