+++ Assets/Player.prefab
@@ -76,7 +76,7 @@ SpriteRenderer:
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: 0
   m_SortingLayer: 0
-  m_SortingOrder: 0
+  m_SortingOrder: 5
   m_Sprite: {fileID: 21300000, guid: f45d2e017eb068f409dc4325dbb0bd7d, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
@@ -185,3 +185,5 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _speed: 2
   _groundMap: {fileID: 0}
+  _isDebug: 0
+  _obj: {fileID: 0}
