--- Assets/Script/CharacterManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02341126dd5aa3e47bc3ec76480d9896
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
