+++ Assets/Script/CharactorManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4837a29a96badad4c9e5c6472500951b
+guid: 02341126dd5aa3e47bc3ec76480d9896
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
